Split command handlers into separate files.
